Remove Java Rhino JS Execution Module
The current source contains a Java module that wraps the javascript execution using Java Rhino. This was originally done as a proof-of-concept, but never went beyond that for a variety of reasons.

We are now removing that module for the following reasons:

- there is now a native Java solution at DBCG/cql_engine
- the Rhino-based solution is no longer maintained and has broken tests
- there are likely better approaches to wrapping JS in Java now anyway, so even as a proof-of-concept, this is now behind the times and not the recommended approach.

See: https://github.com/cqframework/clinical_quality_language/issues/134