Bring in moment.js for DateTime DifferenceBetween
To deal with the daylight savings time spring forward/fall back issue,
I brought in the moment.js library.  We still use the existing logic if
the unitField is year or month, because moment.js does not use the
boundary for those fields as its way (it is instead more like CQL's
duration).

Had to update one test that was assuming a timezone offset that it
shouldn't have been.