Remove use of constructor.name
Using `constructor.name` to determine if an instance is something doesn't work when the code is minified/uglified (as in webpack).

We also often still can't use `instanceof` due to circular import issues.

Instead, add `isX` getters to those classes that we used to check by `constructor.name`.  For example, add an `isQuantity` getter to `Quantity` that returns `true`.  Then we can check `foo.isQuantity` instead of `foo.constructor.name == 'Quantity'`.  Other classes won't have `isQuantity` defined, so the result of `nonQuantity.isQuantity` will be `undefined` (which is falsey).

This is needed to support webpacking cql-execution.