Implement basic function overloading and refactor support for Is
Support function overloading based on number of arguments,
then argument types. There is still a limitation in that it doesn't
yet have an algorithm to choose a function when all functions are
valid against the argument types (in which case, the function w/
"least conversion" should be used).

In the process of implementing this, I noticed that the Context
class already had 95% of what we needed to support the "Is"
operator as well, so I refactored "Is" to use the Context functions.

Also added tests for all these scenarios.