Decaffeinate 2020 is operator with overloading (#146)
* Initial pass at supporting Is expressions. Will fall back to previous behavior (throwing an error) if _is isn't implemented in the data model

* Add _is support through BaseDataType

* refactor Is class a little, and add _is to cql-patient for testing purposes

* first pass at overloading implementation. keeps all functions, uses the first one

* Revert "Add _is support through BaseDataType" (#142)

* Revert "Add _is support through BaseDataType"

This reverts commit 343f0035867b999e98341fb68462982a3055921d. CQL
Datatype _is checking is now supported through matches`NamedTypeSpecifier

* Update cql4browsers

* Prettier fixes

* Implement basic function overloading and refactor support for Is (#141)

Support function overloading based on number of arguments,
then argument types. There is still a limitation in that it doesn't
yet have an algorithm to choose a function when all functions are
valid against the argument types (in which case, the function w/
"least conversion" should be used).

In the process of implementing this, I noticed that the Context
class already had 95% of what we needed to support the "Is"
operator as well, so I refactored "Is" to use the Context functions.

Also added tests for all these scenarios.

* Update _is operator argument to take the type specifier from CQL

* Update cql-patient Record to use new signatures

* Decaf 2020 is operator update as implementation (#145)

* Attempt at 'As' operator. added some nullchecks to function overloading

* Fix broken test by allowing null to match on a NamedTypeSpecifier

Co-authored-by: Chris Hossenlopp <hossenlopp@mitre.org>
Co-authored-by: Chris Moesel <cmoesel@mitre.org>

* Convert operandType to operandTypeSpecifier when necessary

OperandDefs can have an operandTypeSpecifier or an operandType.  If the latter is found, convert it to an operandTypeSpecifier.  In practice, CQL-to-ELM does not seem to ever output operandType -- but this is a safety measure in case it ever does.

Co-authored-by: Chris Hossenlopp <hossenlopp@mitre.org>
Co-authored-by: Chris Moesel <cmoesel@users.noreply.github.com>
Co-authored-by: Chris Moesel <cmoesel@mitre.org>