Fix Interval SameOrAfter/SameOrBefore for Integer and Decimal
The previous implementation used this.end().sameOrAfter(other.start(), p), but sameOrAfter is defined only for certain types, so it failed if the interval was numeric (integer/decimal).  Now we use the relevant comparison functions (which in turn, use sameOrAfter only if the type supports the function -- which is good!).