Use Luxon for Duration, Difference, etc
Simplify all the custom code in our implementation by leveraging Luxon more.  This also fixes a few bugs we had in our own implementation.

NOTE: This removes `toUncertainty` as I don't expect it was used outside of our module -- but it retains `toJSDate` in case anyone else is using it.