Replace ucum lib (#217)
* Quantity equivalence should do unit conversion

* Replace ucum lib w/ ucum-lhc

The ucum lib was quite old and unmaintained. It also had some ugly bits.  So this commit replaces it w/ a UCUM library from NLM.

As part of this, I also extracted all of the unit-based capabilities into their own module.  I also had to reimplement some of the math stuff, particulary getting the correct result units when multiplying and dividing.  I think (hope) the implementation is actually better now than it was.  It is at least more understandable than the previous implementation.