Replace (broken) FHIR model with Simple model & fix broken logic/tests (#137)
* Replace (broken) FHIR model with Simple model & fix broken logic/tests

The primary activity of this commit is to remove the generated FHIR
data model since it was based on an old unofficial version of FHIR and
was very buggy. Implementors should use the cql-exec-fhir project
instead.

In place of the old FHIR data model implementation, a new "Simple"
model has been created for development and testing purposes only. All
tests and examples have been updated to use this Simple model instead.

In the course of implementing this work, I also fixed CQL-to-ELM
translation errors for text fixtures that contained invalid CQL. This
uncovered some logic that actually was not working, but the bad test
cases were obscuring that. So this commit also contains some fixes
related to those discoveries -- mainly in the area of detecting
undeflows and overflows in data and quantity math.

* Resolve lodash to ^4.17.19

Resolve lodash to v4.17.19 or above to address reported vulnerability (and make yarn audit happy).

* Update cql-to-elm (1.4.9) and gradle (6.5.1)

In order to work w/ Java 11, we need to update cql-to-elm and gradle.