Record evaluated records on a patient-by-patient basis
There was a bug that caused the Results.evaluatedRecords array to only contain the evaluated records of the most recently processed patient.

After discussing w/ the abacus team, we decided to switch this to record evaluated records patient by patient.  A new Results.patientEvaluatedRecords property was introduced, which is an object w/ patient ids as keys.

I don't think anyone is using the old Results.evaluatedRecords array, but just in case, I added a getter to provide backward-compatibility.

Surprisingly (and somewhat disturbingly) this did not break any tests.