bumping cql-to-elm to version 1.5.2. s (#239)
* bumping cql-to-elm to version 1.5.2. Fixing model laoding issue in JavaScriptTestDataGenerator introduced with change

* updated deps to dea lwith audit checks

* Support Interval<T> overload of expand

In CQL 1.4, expand took a list of intervals.  CQL 1.5 allows for a single interval.  In this case, we simply convert it to a list for our processing.

* Fix translation of spec-tests

Due to a bug in the CQL 1.5.2 translator, the spec tests can't properly compile w/ the Simple model.  Use the QUICK model instead.  This doesn't actually matter since we don't use any components of the model anyway.

* Oops. Fix cql4browsers

Co-authored-by: Chris Moesel <cmoesel@mitre.org>