Support resolving with path.
When the IG Publisher (using binary adjunct support) converts CQL to ELM it creates a system in the library identifier using the canonical base, and creates a full URI for the path in includes.

Here is the identifiers having `system` specified as a URI:

```json
{
  "library" : {
     "annotation" : [ {
        "translatorOptions" : "EnableAnnotations,EnableLocators,DisableListDemotion,DisableListPromotion",
        "type" : "CqlToElmInfo"
     } ],
     "identifier" : {
        "id" : "HypertensionCaseFeatureLogic",
        "system" : "http://acme.com/canonical-base",
        "version" : "0.1.0"
     },
     …
```

And `includes` has a proper URI built for the `path` portion of include, as seen below:

```json
  "includes": {
    "def": [
      {
        "localId": "2",
        "locator": "5:1-5:54",
        "localIdentifier": "FHIRHelpers",
        "path": "http://hl7.org/fhir/FHIRHelpers",
        "version": "4.0.1",
      ...
```

This patch allows it to work when system is used and the resolve path is a full path.