Use passed in executionDateTime timeZoneOffset when available (#313)
Previously, getDateTime() in dateTime.ts was not passing in a timeZoneOffset when creating a new DateTime. We only want to do this if an executionDateTime is not provided since it will set the timeZoneOffset to the timeZoneOffset of the evaluation. In the case where the executionDateTime is passed in, we want the  timeZoneOffset to come from that. We now achieve this by passing the tx.getExecutionDateTime().timeZoneOffset to functions that call getDateTime().