Add getFunction to the Context (#338)
Fixes #335 by adding a getFunction() method to the Context class which goes up the parent context tree until it gets to the Library and is able to execute getFunction.

This fixes the issue of an overloaded function that has the same name as one of its arguments and calls itself (or another overload). The broad context.get() method was being used to get the possible FunctionDefs was returning the first thing that matches the name of the function. So it was ending up with the value of the argument passed in which had the same name. Tests are included that replicate this issue with multiple modes of execution.

This PR also updates the cql-to-elm version for test data generation to 3.22.0 and updates the translator options to roughly match the options used by MADiE. 

Included (squashed) commits:

* initial adding of test for fluent overloading issue

* updated cql-to-elm to 3.22.0 and added some translator options that MADiE uses

* Added test for overloaded functions calling themselves

* resolved npm audit concern with babel and fixed prettier issue with test-data file

* reworked test to have an argument name matching the overloaded function name. fixed issue by adding getFunction to Context

* add missing newline to package-lock.json that npm insists on making

* added info on translator settings. cql fix for ProperIncludedIn test. re-built test, fixing some post rebase stuff. added new test library to prettier ignore

* renamed new tests. rebuilt spec-test data with new translator and added workaround for ProperIncludedIn null test