Handle valuesets passed into functions (#341)
Partially addresses #226 by further generalizing the building of the valueset fields for AnyInValueSet, InValueSet, and ExpandValueSet. All of these expression classes previously assumed that a valueset would come in as a reference (ValueSetRef), but when a valueset is passed into a function, the valueset must be built from a valueset expression.

This also updates some error messages and types to accurately reflect the state that's expected from the translator.

Squashed commits:

* Handle valuesets passed into functions

* Apply suggestions from code review

Additional tests and ensure existing tests check assertions without await

Co-authored-by: hossenlopp <hossenlopp@mitre.org>

* Make AnyIn and In consistent, update tests and test organization

* additional tests for InValueSet with null code

---------

Co-authored-by: hossenlopp <hossenlopp@mitre.org>