Duration uncertainty (#340)
This PR updates the functionality of durationBetween calculations in order to match the latest descriptions in the CQL Logical Specification described. This PR also fixes #325.

According to the CQL Logical Specification, "Note that uncertainty calculations, just like date and time comparison calculations, consider seconds and milliseconds as a single combined precision with decimal semantics". This means that the duration between calculation should NOT produce an Uncertainty in both cases: when ms are defined and when ms are not defined.

Squashed Commits:

* Duration does not produce Uncertainty for undefined milliseconds

* Fix one lint issue, browserify

* disabled lint line, added seconds to CQL file, fixed test, check that seconds is not null

* Don't treat ms as 0 if unitField is ms

* Use constant DateTime.Unit.MILLISECOND