Exec args race condition fix (#336)
This commit addresses a race condition while evaluating the arguments for an expression. When executing the arguments for a statement the expected behavior would be for the values of any of the ExpressionDefs that were executed as part of the argument execution to be cached for subsequent calls to the same ExpressionDef. Due to the nature of the async processing within javascript, this was not happening. Instead, all of the arguments were being executed via Promise.all which in Node initiates all of the promises at once, adding them to the event loop for processing. This causes potential performance issues if the Arguments being executed rely on the same underlying Expressions as they will each execute those additional expressions instead of the first one cacheing the result and subsequent arguments using those cached results.

This commit addresses that issue by simply serializing the processing of arguments using a for loop. This guarantees that any expressions that were executed by previous arguments will be cached and the values used by any arguments processed there after. Note that using array predicates also caused issues of other sorts which is why a simple for loop was used.

Squashed commits:

* Fixing race condition in execArgs

* running browserify on changes

* CommonLib3 library-test tests added to demonstrate ExpressionRef library context values cache issue.

* formatting fixes

* Modified test to check for underlying condition being addressed. Moving tests from library-tests to reusable as the issue is not related to libraries but related to how arguments are executed

* fising lint errors

* removing await

* updated browser library

* removing unused cql in test data

* removed Common3 library no longer used in tests. Move new test to be with ExpressionRef tests

* fully remove unnecessary test section and removed extra newline in library/data.cql

---------

Co-authored-by: Grassart, Aymeric <Grassaa@labcorp.com>
Co-authored-by: Chris Hossenlopp <hossenlopp@mitre.org>