Fix spec-tests runner
The spec-tests runner was never updated to use the async exec. As a result, it always passed all tests (because, apparently, unresolved promises are always equal). This updates the runner to properly await each exec.

This also skips tests that are now failing due to translator issues or unimplemented CQL features. It also fixes the implementation of includes/included in for the overload with two lists, which was returning the wrong result when the second list was null.