Update Spec Tests and Fix Newly Found Issues (#348)
* Fix spec-tests runner

The spec-tests runner was never updated to use the async exec. As a result, it always passed all tests (because, apparently, unresolved promises are always equal). This updates the runner to properly await each exec.

This also skips tests that are now failing due to translator issues or unimplemented CQL features. It also fixes the implementation of includes/included in for the overload with two lists, which was returning the wrong result when the second list was null.

* Update spec-tests

Update the spec-tests based on the latest versions at cqframework/cql-tests. Also update the generator to more gracefully handle the new <library> tag.

Note that there are now broken tests that will be addressed in a future commit.

* AllTrue should return true when items list is null

* AnyTrue should return false when items list is null

* Skip tests that use Long (unimplemented)

* Skip tests that use Modulo by Quantity

* Skip DateTime arithmetic tests with wrong or potentially wrong expected outputs

* Skip interval operator tests that have wrong expected output

* Skip List operators tests that are wrong, we get wrong, or we haven't implemented

* In and Contains should treat nulls as equal (clarified in CQL 1.5)

* Aggregate queries should not be distinct be default

* Support \$ as literal $ in replaceMatches substitution string

* Remove and re-categorized skipped tests

Remove any skipped tests that are now passing and recategorize any that fail for different reasons.

* When comparing items in lists for containment operations, always treat null items as equal

* rebuild new tests on translator 3.27.0