First Class ValueSet (#349)
Addresses #226 by creating a new type (`CQLValueSet`) that aligns with the CQL 1.5 specification defined ValueSet type.

Note: This type naming is currently `CQLValueSet` so as not to change the interface of `ValueSet`, but this is intended to be a temporary decision that should be updated for the next major update. At the next major update, change `CQLValueSet` -> `ValueSet` and `ValueSet` -> `ValueSetExpansion`.

This commit shifts valueset resolution to the point that it is required, such as when `hasMatch` or `expand` is called. It does not cache valueset expansions in the context and leaves valueset caching up to the code service. This includes an update to equivalence that only expands valuesets to check matching codes if the valuesets do not have an exact match of id/version. The commit also adds a consistent error if passed valuesets are not able to be resolved by the code service.

Co-authored-by: Elsa <eperelli@mitre.org>
Co-authored-by: Chris Moesel <cmoesel@mitre.org>
Co-authored-by: hossenlopp <hossenlopp@mitre.org>